Parse and/or sequences
Precedence is currently an emergent behaviour.

(You ought to be able to chain multiple `chainl1` parses
at different levels of precedence, but I don't understand
it and is rather poorly documented, so leaving for now.)

Parens would disambiguate... except the `parens` parser
as implemented works only with simple expressions (e.g.
not with `chainl1` sequences.)

Again, leaving for now, as this gives us something to
play with for now.