Data.Argonaut.Passthrough
TaskWarrior's JSON format is documented at
https://taskwarrior.org/docs/design/task.html

Note the section

> # Additional Attributes
> There MAY be other fields than those listed above in a task
> definition. Such fields MUST be preserved intact by any client, which
> means that if a task is downloaded that contains an unrecognized field,
> that field MUST not be modified, and MUST continue to exist in the
> task..

This means that the strict Codec based approach we're using
doesn't work, as we lose any fields that we've not explicitly listed.

This approach, as discussed with @thomashoneyman on Slack, stores the
entire JSON in an additional field, and merges it back on encode.